fixed range generation, moved test-folder to correct package
