+++ openvalidation-rest/src/test/java/io/openvalidation/test/rest/model/dto/dto/astDTO/transformation/TreeTransformerTest.java
@@ -1,4 +1,4 @@
-package org.bag.openvalidation.test.rest.model.dto.astDTO.transformation;
+package io.openvalidation.test.rest.model.dto.dto.astDTO.transformation;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
